--- a/synedrion/src/cggmp21/params.rs
+++ b/synedrion/src/cggmp21/params.rs
@@ -5,6 +5,7 @@ use core::{fmt::Debug, ops::Add};
 // So as long as that is the case, `k256` `Uint` is separate
 // from the one used throughout the crate.
 use crypto_bigint::{BitOps, NonZero, Uint, U1024, U2048, U4096, U512, U8192};
+use crypto_bigint::{U128, U256};
 use digest::generic_array::{ArrayLength, GenericArray};
 use ecdsa::hazmat::{DigestPrimitive, SignPrimitive, VerifyPrimitive};
 use primeorder::elliptic_curve::{
@@ -15,12 +16,12 @@ use primeorder::elliptic_curve::{
 };
 use serde::{Deserialize, Serialize};
 
-use tiny_curve::TinyCurve64;
+use tiny_curve::TinyCurve32;
 
 use crate::{
     paillier::PaillierParams,
     tools::hashing::HashableType,
-    uint::{U1024Mod, U2048Mod, U4096Mod, U512Mod},
+    uint::{U1024Mod, U128Mod, U2048Mod, U256Mod, U4096Mod, U512Mod},
 };
 
 #[derive(Debug, Copy, Clone, PartialEq, Eq, Serialize, Deserialize)]
@@ -85,14 +86,14 @@ impl PaillierParams for PaillierTest {
     are much smaller than 2*PRIME_BITS.
     */
 
-    const PRIME_BITS: u32 = 397;
-    type HalfUint = U512;
-    type HalfUintMod = U512Mod;
-    type Uint = U1024;
-    type UintMod = U1024Mod;
-    type WideUint = U2048;
-    type WideUintMod = U2048Mod;
-    type ExtraWideUint = U4096;
+    const PRIME_BITS: u32 = 128;
+    type HalfUint = U128;
+    type HalfUintMod = U128Mod;
+    type Uint = U256;
+    type UintMod = U256Mod;
+    type WideUint = U512;
+    type WideUintMod = U512Mod;
+    type ExtraWideUint = U1024;
 }
 
 /// Paillier parameters corresponding to 112 bits of security.
@@ -192,15 +193,17 @@ pub struct TestParams;
 // - Range checks will fail with the probability $q / 2^\eps$, so $\eps$ should be large enough.
 // - P^{fac} assumes $N ~ 2^{4 \ell + 2 \eps}$
 impl SchemeParams for TestParams {
-    type Curve = TinyCurve64;
+    type Curve = TinyCurve32;
+    // TODO: ReprUint is typenum::U192 because of RustCrypto stack internals, hence the U384 here,
+    // but once that is solved, this can be a U128 (or even smaller).
     type WideCurveUint = bigintv05::U384;
     // TODO: 8*24 = 192, this is to work around an issue with the ModulusSize-trait. This should be ideally be 8 bytes long.
     type HashOutput = [u8; 24];
     const SECURITY_BITS: usize = 16;
-    const SECURITY_PARAMETER: usize = 10;
-    const L_BOUND: u32 = 256;
-    const LP_BOUND: u32 = 256;
-    const EPS_BOUND: u32 = 320;
+    const SECURITY_PARAMETER: usize = 32;
+    const L_BOUND: u32 = 32;
+    const EPS_BOUND: u32 = 64;
+    const LP_BOUND: u32 = 160;
     type Paillier = PaillierTest;
     const CURVE_ORDER: NonZero<<Self::Paillier as PaillierParams>::Uint> =
         convert_uint(upcast_uint(Self::Curve::ORDER))
@@ -223,8 +226,8 @@ impl SchemeParams for ProductionParams112 {
     const SECURITY_BITS: usize = 112;
     const SECURITY_PARAMETER: usize = 256;
     const L_BOUND: u32 = 256;
-    const LP_BOUND: u32 = Self::L_BOUND * 5;
     const EPS_BOUND: u32 = Self::L_BOUND * 2;
+    const LP_BOUND: u32 = Self::L_BOUND * 5;
     type Paillier = PaillierProduction112;
     const CURVE_ORDER: NonZero<<Self::Paillier as PaillierParams>::Uint> =
         convert_uint(upcast_uint(Self::Curve::ORDER))
@@ -240,7 +243,7 @@ impl SchemeParams for ProductionParams112 {
 mod tests {
     use super::{
         bigintv05::{U256, U64},
-        upcast_uint, ProductionParams112, SchemeParams,
+        upcast_uint, ProductionParams112, SchemeParams, TestParams,
     };
 
     #[test]
@@ -269,7 +272,6 @@ mod tests {
     #[test]
     fn parameter_consistency() {
         assert!(ProductionParams112::are_self_consistent());
-        // TODO: These are not consistent right now.
-        // assert!(TestParams::are_self_consistent());
+        assert!(TestParams::are_self_consistent());
     }
 }