Adjust test params to be as tight as possible:

- use TinyCurve32
- `L_BOUND` is 32
- `EPS_BOUND` is 64 (2*32)
- `LP_BOUND` is 160 (5*32)
- `PRIME_BITS` is 128 (so Uint is U256 etc)

The speedup when running tests is dramatic.

Master:

```
hyperfine -w 2 "cargo t --release"
 Benchmark 1: cargo t --release
   Time (mean ± σ):      7.790 s ±  0.122 s    [User: 43.535 s, System: 0.446 s]
   Range (min … max):    7.632 s …  8.011 s    10 runs
```

This PR:

```
hyperfine -w 2 "cargo t --release"
Benchmark 1: cargo t --release
  Time (mean ± σ):     942.8 ms ±  18.9 ms    [User: 2882.5 ms, System: 166.0 ms]
  Range (min … max):   919.8 ms … 971.8 ms    10 runs
```

**NOTE**: There is one potentially problematic change in this PR, to be reviewed carefully. The `checked_add` impl for `SecretSigned<T>` used to pick the biggest bound from the two operands and *then add 1* to it. This leads to a few test failures with the params as configured in this PR. The underlying reason is that now the curve order is so much closer to uint sizes that we run into corner cases where e.g. [Paillier decryption yields a SecretSigned with a bound of 255](https://github.com/entropyxyz/synedrion/blob/dp-use-TinyCurve32-in-tests/synedrion/src/paillier/encryption.rs#L224) (which comes from `MODULUS_BITS`), causing addition to always "overflow" (actually: fail the bounds check that `bound + 1 < T::BITS`).

~~We have discussed this bound increment last summer and I think we decided to keep the increment as the conservative choice, but I actually think it's wrong. Maybe now is the right time to make the final call on this?~~

I'm wrong, the last bit must be preserved. However, the simple bounds check breaks down in this corner case and we need to do something smarter for this to work. How about checking the last bit before and after the sum, and if the bit flipped we overflowed?