@fjarri 2025-03-06
That's strange, where exactly in the protocol does the overflow occur? There shouldn't be any problems as long as the modulus size is larger than, roughly `q^2 * n` where `q` is the curve order and `n` is the number of nodes, and we're still way beyond that even with test parameters. It might be that we're actually catching a bug in the logic sopmewhere.