@dvdplm 2025-03-06
> where exactly in the protocol does the overflow occur?

So first of all, let me be clear about what "overflow" means here: it's actually not that the case that there's arithmetic that *actually* overflows with these new params. What happens is that the bounds check in `impl Add<SecretSigned<T>> for SecretSigned<T>` checks that the biggest bound of the operands **plus one** is smaller than `T::BITS`. 
Before this PR, `T::BITS` was 2048 and the numbers involved were bounded on the curve modulus or `LP_BOUND`/`EPS_BOUND` so much much smaller. With this PR, the uint sizes used and the curve order (which determines `MODULUS_BITS`, `LP_BOUND` and `EPS_BOUND`) are much closer so the "coarse" bounds check now matters much more.

Take [`r5_dec_failed`](https://github.com/entropyxyz/synedrion/blob/dp-use-TinyCurve32-in-tests/synedrion/src/cggmp21/interactive_signing_tests.rs#L511) for instance. There's an "overflow" when adding `beta` and `secret.y` [here](https://github.com/entropyxyz/synedrion/blob/dp-use-TinyCurve32-in-tests/synedrion/src/cggmp21/sigma/dec.rs#L151), because `secret.y` has a [bound of 255](https://github.com/entropyxyz/synedrion/blob/dp-use-TinyCurve32-in-tests/synedrion/src/paillier/encryption.rs#L224) (`beta` has a bound of `LP_BOUND + EPS_BOUND = 224` but that doesn't matter here). The "overflow" happens [because `255 + 1` is not less than `T::BITS` (256)](https://github.com/entropyxyz/synedrion/blob/master/synedrion/src/uint/secret_signed.rs#L207).

When I say that the current "overflow" check is too coarse, what I mean is a `SecretSigned<U256>` with bound 255 will always overflow, regardless of what numbers are involved. Is that ok? Why not actually check if there's an overflow, so that `SecretSigned::from(value: U256::ZERO, bound: 255) + SecretSigned::from(value: U256::ONE, bound: 123)` can work (pseudocode)?