@fjarri 2025-03-06
But that's specifically #187. It's not caused by the bound propagation, there still should be plenty of space. It's a drawback of the scheme which should be resolved on the "theoretical" level, then the implementation level will be correct. 

As for what to do with it right now, I'm not sure - need to think about it.