@dvdplm 2025-03-07
> As for what to do with it right now, I'm not sure

Seems like we shouldn't rush this. I'll back out my changes to `checked_add` and loosen the params to make it work. That way we can merge this and enjoy the fast tests while we ponder what we do here.